+++ lib/chain/tezos/tezos_node_writer.dart
@@ -203,6 +203,7 @@ class TezosNodeWriter {
           'value': jsonDecode(michelineLambda)
         };
       }
+      
     } else if (entrypoint != null) {
       transaction.parameters = {'entrypoint': entrypoint, 'value': []};
     }
@@ -276,8 +277,6 @@ class TezosNodeWriter {
         'signature': signedOpGroup['signature']
       }
     ];
-    print("signedOpGroup['signature'] ===> ${signedOpGroup['signature']}");
-    print("parameters ===> ${jsonEncode(payload)}");
     var response = await HttpHelper.performPostRequest(server,
         'chains/$chainid/blocks/head/helpers/preapply/operations', payload);
     var json;
@@ -358,7 +357,6 @@ class TezosNodeWriter {
       int counter) {
     var parsedCode;
     var parsedStorage;
-    print("using code Format ==> ${codeFormat.toString()}");
     if (codeFormat == TezosParameterFormat.Michelson) {
       parsedCode =
           jsonDecode(TezosLanguageUtil.translateMichelsonToMicheline(code));