+++ lib/michelson_parser/parser/nearley.dart
@@ -367,7 +367,7 @@ class ColumnState {
   void finish() {
     if (rule.postprocess != null) {
       try {
-        data = rule.postprocess(data);
+        data = rule.postprocess(data);  
       } catch (e) {
         print("Error In ===> " + rule.name);
       }