+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [2.1.2-beta.1]
+
+* Operation Gas fee calculator optimisation
+* Michelson grammar improvements
+* Minor bug fixes
+
 # [2.1.2-beta]
 
 * Support for florence mainnet and testnet